+++ hs-bindgen/src-internal/HsBindgen/Frontend.hs
@@ -166,7 +166,8 @@ import HsBindgen.Util.Tracer
 --   that the unusable declaration /and all of its dependencies/ will not be
 --   selected.
 runFrontend ::
-     Tracer FrontendMsg
+     HasCallStack
+  => Tracer FrontendMsg
   -> FrontendConfig
   -> BootArtefact
   -> IO FrontendArtefact
@@ -212,7 +213,7 @@ runFrontend tracer config boot = do
       afterParse <- parsePass
       let (afterSimplifyAST, msgsSimplifyAST) =
             simplifyAST afterParse.usageAnalysis afterParse.results
-      forM_ msgsSimplifyAST $ traceWith tracer . FrontendSimplifyAST
+      forM_ msgsSimplifyAST $ traceWith tracer . extendCallStackMsg FrontendSimplifyAST
       pure afterSimplifyAST
 
     assignAnonIdsPass <- cache "assignAnonIds" $ do
@@ -220,7 +221,7 @@ runFrontend tracer config boot = do
       afterSimplifyAST <- simplifyASTPass
       let (afterAssignAnonIds, msgsAssignAnonIds) =
             assignAnonIds afterParse.usageAnalysis afterSimplifyAST
-      forM_ msgsAssignAnonIds $ traceWith tracer . FrontendAssignAnonIds
+      forM_ msgsAssignAnonIds $ traceWith tracer . extendCallStackMsg FrontendAssignAnonIds
       pure afterAssignAnonIds
 
     constructTranslationUnitPass <- cache "constructTranslationUnit" $ do
@@ -235,7 +236,7 @@ runFrontend tracer config boot = do
       std <- boot.cStandard
       let (afterHandleMacros, msgsHandleMacros) =
             handleMacros std afterConstructTranslationUnit
-      forM_ msgsHandleMacros $ traceWith tracer . FrontendHandleMacros
+      forM_ msgsHandleMacros $ traceWith tracer . extendCallStackMsg FrontendHandleMacros
       pure afterHandleMacros
 
     resolveBindingSpecsPass <- cache "resolveBindingSpecs" $ do
@@ -248,21 +249,21 @@ runFrontend tracer config boot = do
               extSpecs
               pSpec
               afterHandleMacros
-      forM_ msgsResolveBindingSpecs $ traceWith tracer . FrontendResolveBindingSpecs
+      forM_ msgsResolveBindingSpecs $ traceWith tracer . extendCallStackMsg FrontendResolveBindingSpecs
       pure afterResolveBindingSpecs
 
     mangleNamesPass <- cache "mangleNames" $ do
       afterResolveBindingSpecs <- resolveBindingSpecsPass
       let (afterMangleNames, msgsMangleNames) =
             mangleNames config.fieldNamingStrategy afterResolveBindingSpecs
-      forM_ msgsMangleNames $ traceWith tracer . FrontendMangleNames
+      forM_ msgsMangleNames $ traceWith tracer . extendCallStackMsg FrontendMangleNames
       pure afterMangleNames
 
     adjustTypesPass <- cache "AdjustTypes" $ do
       afterMangleNamesPass <- mangleNamesPass
       let (afterAdjustTypes, msgsAdjustTypes) =
             adjustTypes afterMangleNamesPass
-      forM_ msgsAdjustTypes $ traceWith tracer . FrontendAdjustTypes
+      forM_ msgsAdjustTypes $ traceWith tracer . extendCallStackMsg FrontendAdjustTypes
       pure afterAdjustTypes
 
     selectPass <- cache "select" $ do
@@ -274,7 +275,7 @@ runFrontend tracer config boot = do
               afterParse.isInMainHeaderDir
               selectConfig
               afterAdjustTypesPass
-      forM_ msgsSelect $ traceWith tracer . FrontendSelect
+      forM_ msgsSelect $ traceWith tracer . extendCallStackMsg FrontendSelect
       pure afterSelect
 
     finalPass <- cache "Final" $ selectPass